chore: component separation
