+++ apps/web/src/services/info/components/NoticeForm.vue
@@ -1,37 +1,21 @@
 <script setup lang="ts">
 import { computed, reactive, watch } from 'vue';
-import { useRouter } from 'vue-router/composables';
 
 import {
     PButton,
     PCheckbox,
     PDataLoader,
-    PEmpty,
     PFieldGroup,
     PPaneLayout, PRadio, PRadioGroup,
-    PSelectDropdown,
-    PTextHighlighting,
     PTextInput,
-    PBadge,
 } from '@spaceone/design-system';
-import type {
-    AutocompleteHandler,
-    SelectDropdownMenuItem,
-} from '@spaceone/design-system/types/inputs/dropdown/select-dropdown/type';
-
-import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
-import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
 import { SpaceRouter } from '@/router';
-import type { ListResponse } from '@/schema/_common/api-verbs/list';
 import type { PostUpdateParameters } from '@/schema/board/post/api-verbs/update';
 import { POST_BOARD_TYPE } from '@/schema/board/post/constant';
-import type { WorkspaceListParameters } from '@/schema/identity/workspace/api-verbs/list';
 import type { WorkspaceModel } from '@/schema/identity/workspace/model';
 import { store } from '@/store';
 import { i18n } from '@/translations';
-
-import { makeAdminRouteName } from '@/router/helpers/route-helper';
 
 import { useUserWorkspaceStore } from '@/store/app-context/workspace/user-workspace-store';
 
@@ -44,17 +28,12 @@ import ErrorHandler from '@/common/composables/error/errorHandler';
 import { useFileUploader } from '@/common/composables/file-uploader';
 import { useFormValidator } from '@/common/composables/form-validator';
 import { useProperRouteLocation } from '@/common/composables/proper-route-location';
-import WorkspaceLogoIcon from '@/common/modules/navigations/top-bar/modules/top-bar-header/WorkspaceLogoIcon.vue';
 
+import NoticeWorkspaceDropdown from '@/services/info/components/NoticeWorkspaceDropdown.vue';
 import { INFO_ROUTE } from '@/services/info/routes/route-constant';
 import { useNoticeDetailStore } from '@/services/info/stores/notice-detail-store';
-import { PREFERENCE_ROUTE } from '@/services/preference/routes/route-constant';
+import type { WorkspaceDropdownMenuItem } from '@/services/info/types/notice-type';
 
-interface WorkspaceDropdownMenuItem extends SelectDropdownMenuItem {
-    tags?: {
-        theme?: string;
-    };
-}
 interface Props {
     type?: NoticeFormType;
 }
@@ -69,8 +48,6 @@ const noticeDetailState = noticeDetailStore.state;
 const userWorkspaceStore = useUserWorkspaceStore();
 const userWorkspaceGetters = userWorkspaceStore.getters;
 
-const router = useRouter();
-
 const storeState = reactive({
     workspaceList: computed<WorkspaceModel[]>(() => userWorkspaceGetters.workspaceList),
 });
@@ -81,11 +58,7 @@ const state = reactive({
     attachments: [] as Attachment[],
 });
 const workspaceState = reactive({
-    loading: true,
-    visible: false,
-    menuItems: [] as WorkspaceDropdownMenuItem[],
     selectedItems: [] as WorkspaceDropdownMenuItem[],
-    searchText: '',
     radioMenuList: computed(() => ([
         i18n.t('INFO.NOTICE.FORM.ALL'),
         i18n.t('INFO.NOTICE.FORM.SPECIFIC_WORKSPACE'),
@@ -129,12 +102,7 @@ const formData = computed<Omit<PostUpdateParameters, 'post_id'>>(() => ({
 const { fileUploader } = useFileUploader();
 const { getProperRouteLocation } = useProperRouteLocation();
 
-const workspaceMenuHandler: AutocompleteHandler = async (inputText: string) => {
-    await fetchListWorkspaces(inputText);
-    return {
-        results: workspaceState.menuItems as WorkspaceDropdownMenuItem[],
-    };
-};
+
 const handleConfirm = () => {
     if (props.type === 'CREATE') handleCreateNotice();
     else if (props.type === 'EDIT') handleEditNotice();
@@ -175,32 +143,6 @@ const handleEditNotice = async () => {
     }
 };
 
-const workspaceListApiQueryHelper = new ApiQueryHelper()
-    .setPageStart(1).setPageLimit(15)
-    .setSort('name', true);
-const fetchListWorkspaces = async (inputText: string) => {
-    workspaceState.loading = true;
-
-    workspaceListApiQueryHelper.setFilters([
-        { k: 'name', v: inputText, o: '' },
-        { k: 'state', v: 'ENABLED', o: '' },
-    ]);
-    try {
-        const { results } = await SpaceConnector.clientV2.identity.workspace.list<WorkspaceListParameters, ListResponse<WorkspaceModel>>({
-            query: workspaceListApiQueryHelper.data,
-        });
-        workspaceState.menuItems = (results ?? []).map((role) => ({
-            label: role.name,
-            name: role.workspace_id,
-            tags: role.tags,
-        }));
-    } catch (e) {
-        ErrorHandler.handleError(e);
-    } finally {
-        workspaceState.loading = false;
-    }
-};
-
 watch([() => noticeDetailState.post, () => noticeDetailState.loading], async ([notice, loading]) => {
     if (loading) return;
 
@@ -217,12 +159,12 @@ watch([() => noticeDetailState.post, () => noticeDetailState.loading], async ([n
         workspaceState.selectedItems = [];
     } else {
         workspaceState.selectedRadioIdx = 1;
-        await fetchListWorkspaces('');
         workspaceState.selectedItems = (notice?.workspaces ?? []).map((workspace) => {
             const selectedWorkspace = storeState.workspaceList.find((w) => w.workspace_id === workspace);
             return {
                 label: selectedWorkspace?.name || '',
                 name: selectedWorkspace?.workspace_id || '',
+                tags: selectedWorkspace?.tags,
             };
         });
     }
@@ -263,79 +205,9 @@ watch([() => noticeDetailState.post, () => noticeDetailState.loading], async ([n
                             </span>
                         </p-radio>
                     </p-radio-group>
-                    <p-select-dropdown v-if="workspaceState.selectedRadioIdx === 1"
-                                       use-fixed-menu-style
-                                       :visible-menu.sync="workspaceState.visible"
-                                       :loading="workspaceState.loading"
-                                       :search-text.sync="workspaceState.searchText"
-                                       :selected.sync="workspaceState.selectedItems"
-                                       :handler="workspaceMenuHandler"
-                                       show-select-marker
-                                       is-filterable
-                                       multi-selectable
-                                       appearance-type="badge"
-                                       show-delete-all-button
-                                       class="workspace-select-dropdown"
-                                       :class="{'no-data': workspaceState.menuItems.length === 0 && !workspaceState.loading}"
-                    >
-                        <template #dropdown-button>
-                            <div v-if="workspaceState.selectedItems.length > 0"
-                                 class="selected-workspace-wrapper"
-                            >
-                                <workspace-logo-icon :text="workspaceState.selectedItems[0].label || ''"
-                                                     :theme="workspaceState.selectedItems[0].tags?.theme"
-                                                     size="xxs"
-                                />
-                                <span>{{ workspaceState.selectedItems[0].label }}</span>
-                                <p-badge v-if="workspaceState.selectedItems.length > 1"
-                                         style-type="blue200"
-                                         badge-type="subtle"
-                                >
-                                    + {{ workspaceState.selectedItems.length - 1 }}
-                                </p-badge>
-                            </div>
-                            <span v-else
-                                  class="placeholder"
-                            >
-                                {{ $t('INFO.NOTICE.FORM.SELECT') }}
-                            </span>
-                        </template>
-                        <template #menu-item--format="{item}">
-                            <div class="menu-item-wrapper">
-                                <workspace-logo-icon :text="item?.label || ''"
-                                                     :theme="item?.tags?.theme"
-                                                     size="xxs"
-                                />
-                                <p-text-highlighting class="label-text"
-                                                     :text="item.label"
-                                                     :term="workspaceState.searchText"
-                                />
-                            </div>
-                        </template>
-                        <template #no-data-area>
-                            <p-empty v-if="workspaceState.menuItems.length === 0 && !workspaceState.loading"
-                                     image-size="sm"
-                                     show-image
-                                     show-button
-                                     class="no-data-wrapper"
-                            >
-                                <template #image>
-                                    <img src="@/assets/images/illust_planet.svg"
-                                         alt="empty-options"
-                                    >
-                                </template>
-                                <template #button>
-                                    <p-button style-type="substitutive"
-                                              icon-left="ic_plus_bold"
-                                              @click="router.push({ name: makeAdminRouteName(PREFERENCE_ROUTE.WORKSPACES._NAME) })"
-                                    >
-                                        {{ $t('INFO.NOTICE.FORM.CREATE_WORKSPACE') }}
-                                    </p-button>
-                                </template>
-                                {{ $t('INFO.NOTICE.FORM.NO_WORKSPACE') }}
-                            </p-empty>
-                        </template>
-                    </p-select-dropdown>
+                    <notice-workspace-dropdown :selected-radio-idx="workspaceState.selectedRadioIdx"
+                                               :selected-items.sync="workspaceState.selectedItems"
+                    />
                 </p-field-group>
                 <p-field-group class="notice-label-wrapper"
                                :label="$t('INFO.NOTICE.FORM.LABEL_TITLE')"
@@ -416,26 +288,6 @@ watch([() => noticeDetailState.post, () => noticeDetailState.loading], async ([n
     }
     .notice-create-options-wrapper {
         @apply flex flex-col gap-2;
-    }
-
-    .workspace-select-dropdown {
-        width: 50%;
-        margin-top: 0.25rem;
-        .no-data-wrapper {
-            margin-top: 2rem;
-            margin-bottom: 2rem;
-        }
-        .menu-item-wrapper {
-            @apply flex items-center;
-            gap: 0.25rem;
-        }
-        .selected-workspace-wrapper {
-            @apply flex items-center;
-            gap: 0.25rem;
-        }
-        .placeholder {
-            @apply text-gray-600;
-        }
     }
 }
 .notice-create-buttons-wrapper {
