+++ apps/web/src/services/info/types/notice-type.ts
@@ -1,6 +1,14 @@
 import type { TranslateResult } from 'vue-i18n';
 
+import type { SelectDropdownMenuItem } from '@spaceone/design-system/types/inputs/dropdown/select-dropdown/type';
+
 export interface NoticePostBadgeInfo {
     label: TranslateResult;
     style: string;
+}
+
+export interface WorkspaceDropdownMenuItem extends SelectDropdownMenuItem {
+    tags?: {
+        theme?: string;
+    };
 }
